Update achievements list test, add achievement
updated the achievements list test. the only thing i can think of that
will never change is that the achievement should always show the name,
so that's what i tested for.

added the achievement test, and the only requirement there was that on-
click the achievement should show the tooltip if it's unachieved. that's
what the test checks for.

added an achievement component which passes this test.